requirements.txt
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
lxml==4.6.2
# pkg-resources==0.0.0
premailer==3.7.0
python-decouple==3.3
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
